Cosmetics
Moved Hex4 declarations from KML into Helper.
